Generalize SHP downloader
